buf: Compile with 'extra' and 'error'.
Includes a fix for the following:
FAILED: clang -Wall -Wextra -Werror -MMD -MF out/sbuf.o.d -c sbuf.c -o
out/sbuf.o
sbuf.c:65:18: error: comparison of integers of different signs: 'int'
and 'size_t' (aka 'unsigned long') [-Werror,-Wsign-compare]
        if (want > b->cap) {
            ~~~~ ^ ~~~~~~

+++ buf/sbuf.c
@@ -59,7 +59,7 @@ void sbuf_free(sbuf *b)
 
 void sbuf_grow(sbuf *b, size_t extra)
 {
-	int want;
+	size_t want;
 
         want = b->len + extra;
         if (want > b->cap) {
